move extensible credential into separate file
